remove local mode
